lkl: add ipv4 encapsulation for raw socket virtio backend
This virtio-net backend is specially useful when we use on Android phone
over cellular interface.  In this use case, the IP address for a virtual
NIC (in LKL) should use the same IP address of the host's one (i.e.,
rmnet0).

IPv6 encapsulation should also support with this mode but it's not
currently implemented.

Signed-off-by: Hajime Tazaki <thehajime@gmail.com>